Page to edit account customer done and registration is so dead
The edit account page NEEDS some CSS refactoring. it's working but the style is NOT it.

The registration:
I added a getCustomer method to then put that into the registration for POST.
Also make a registerToSession that should work.... however
I am unable to make handle the getCustomer make the HTTP request before the rest of the registerToSession method runs.